Add improved energy breakdown summaries
Track and print energy breakdown for different on-chip components
for synapse, dendrite, soma and network h/w. I found this useful
during experiments and exploration.

Also add in some framework to start probing the energy and latency
of individual h/w units. Again this info is readily available but
I just need a mechanism to store in the traces. Other simulators
have a flexible mechanism for tracking whatever metrics, so maybe
I need to create something to track additional metrics, that the
user can easily enable or disable. In this case though, we want
per-tile, per-core and per-unit energy/latency breakdowns