Add detailed breakdowns of on-chip energy
Support a new set of optional performance traces that let the user
toggle on/off tile, core, and unit-level energy metrics. Latency
is tentatively added but not fully supported yet, in theory this
could help calculate utilization.

In the future, if it's useful, I could add even more performance
traces. For example, I could store how many neurons fired, spikes,
packets sent and so on for each tile and core.