Support Booksim2 for cycle-accurate NoC simulation
Using a custom fork of Booksim2, I already supported some level of
integration with the NoC simulator. This commit makes a Booksim2 a
project dependency, allowing CMake to fetch and build, and link
the simulator to SANA-FE. Still just preliminary support with more
to do, currently we're stuck with a fixed file for passing message
info.

TODO: work across the Booksim2 repo to support passing of message
timing info e.g., as a vector, and integrate this feature here,
avoiding the need for file based communciation.
TODO: pass the config file as a string, which can in future be
dynamically generated. This avoids the need for multiple different
config files - ideally everything goes in the arch description