Store booksim config as a ptr for plugin build
Avoid including the booksim libraries in chip.hpp, since it messes
up the plugin build (which needs to include the chip.hpp structures)
Do this by simply using a unique_ptr to dynamically store the config.

This has another advantage that if, at a later point I want to
disable booksim initialization (e.g., if we don't want to enable it
for any runs), maybe I can save on creating a large config struct
every time. This might reduce the size of the SpikingChip struct