Support saving SNN as netlist again
In a bunch of experiments I found it pretty useful to be able to
save generated networks. Re-enable this feature again for netlists,
with support for saving scalar model parameters as strings.

TODO: support saving to YAML files. As part of this, I'll need to
 make it possible to save vectors of model attributes
TODO: enforce mapping order better