Support saving YAML network files
Now we support dumping of SNNs as either YAML or netlist files.
Functional, but can be verbose - currently I don't compress range
of neurons or mappings, and there is a minor bug where it seems
to redefine common group attributes.

Also, hyperedges are currently not supported. These are all nice to
have features for the future, if needed.