Minor changes to spiking hw structs
Tried profiling the message processing and making some improvements.
We duplicate the message processing pipeline a bunch in each core,
although these vectors are tiny (just 1-3 elements). I found that
actually it's worse if we remove these local copies of the pipeline
in each connection.. so leave these be. Don't both with keeping a
single global copy of the processing pipelines.