Clean and thus optimize pipeline input process()
Before a bunch of computation was being done before every pipeline
call unnecessarily. Better structure and optimize this code, so
that we initialize function pointers at creation and then simply
call the correct input/output interfaces for each pipeline h/w unit.
This reduces a lot of branching and code execution during neuron
and message processing. This improves the performance of both
processing blocks by ~20%. As a bonus, the code is cleaner and
better encapsulated.