[wip] More improvements to multithreaded scheduler
Also implement the writing of messages to trace in order and
calculation of total latency. Wrap the lists of shared messages
with a shared_ptr - this is because before I was copying the
messages over to the scheduler structures, which is relatively
expensive. A decent chunk of time was being spent pushing these
structs (comparable to the total neuron processing). Now that
we just use a shared ptr, this cost is basically negligible.

With these latest scheduler and processing improvements

However, the code is a little messy/hacky and needs to be cleaned
up. I especially need to consider how we can just have a single
logical path for processing messages.

TODO: clean up!
TODO: thinking about it, we probably don't run long enough to get
 dynamic performance info. Instead, I think we can let the user
 select the number of threads for processing and scheduling
 manually. Otherwise, we calculate these based on a fixed scheme
 e.g., just allocate one separate thread