Clean up multithreaded scheduler
Found a cleaner solution and removed some redundant run
data summarizing. Now we just have one logical path for
all schedulers and different numbers of threads. Always
push timesteps to the write queue and consume later. This
is minimal overhead (when scheduling on main thread) but
it simplifies the code