loihi.yaml: Disable noise stream
Disable the noise stream which is a pointer to a specific file
in my local env. I need a better solution if I want this to
be in the public file.

TODO: Either remove the stochastic soma or
create an LFSR implementation that is good enough but just
not bit-perfect with Loihi.