Add checking for supported attributes
Keep a list of all supported attributes for each model and throw
a warning if we try and set an unsupported attribute. This way
we should flag if we have typos in the SNN or arch descriptions.

If we fail to set an attribute (and this matters), that can
already be detected inside the model implementation