Reenable access to groups and neuron containers
It seems it will be useful to still get access to
SpikingNetwork::groups as a Dict or NeurongGroup::neurons
as a list. This allows the user to do their own
manipulations, and get a list of all group names easily.

This feature was only just removed, so simply add it back...
This time, make sure both a readonly properties