Major refactor and reshuffle of files
Create a bunch of new files and break off code from the huge file
chip.cpp. This is kind of backtracking on a change I made a while
ago, but after code review it is clear that a 2000+ file with a ton
of code is not going to be manageable going forward. The problem
before was it wasn't always clear what should live where, but I
think I've found a way of encapsulating code while still keeping
things clear.