Fix message receive delay not being set
This is a bug that was introduced recently by all the code
changes and refactors.

The list of received messages into each core was only being stored
as a copy, meaning that we weren't updating messages correctly
with their receive delays. Make this list a set of references
instead so messages are correctly tracked.