Fix bug with accessing mapped neurons
Mapped neurons were being correctly stored somehow in a vector that
was being accessed with [] indexing, despite that vector only being
allocated with reserve() but not resize(). When trying to access
these using the Python API, the vectors were coming back with size 0

The fix was to split functionality into a new routine that updates
mapped neuron tracking, that populates the vector of references
and then sorts it by offset order. This bug was introduced because
I switched to using reference_wrappers in this vector instead of
raw pointers