Support adding a fixed delay per timestep
I observed that on Loihi there is a fixed delay every timestep,
required for synchronization and assuming some housekeeping by the
CPUs on the chip. This delay is significant for some of the
applications I run (especially DVS gesture). Allow a user to
specify a fixed delay that gets added every timestep.

This fixed delay could also be useful in other architectures e.g.,
that use a fixed clock period to determine the latency.