Implement latency model for timestep sync
Create a simple table-based model for the end of time-step sync
cost. The user can provide nothing (assumed 0), a fixed rate
(e.g., a clock period) or a more complex table of
<tile count, sync delay> pairs.

For Loihi I observed that the synchronization delay (including CPU
housekeeping) varies depending on the number of tiles. This
is enough to significantly impact results (1-6% MAPE difference
across different benchmarks). For some benchmarks, the value is
fixed, but for randomly generated SNNs, we range from 1-32 tiles.
This 0.5-1.7us adjustment will help improve most of our latency
predictions to be within 10% MAPE, which was a benchmark I was
aiming for.