Disable all prints for certain builds
Make the debug build default, but for release builds we want reduced
prints. For the standalone simulator, strip the file info from all
prints. For the Python library, don't print anything to stdout
when using the Release build.

Mark a few variables [[maybe_unused]] that are only used for
prints