[wip] Improve Python simulation interface
Make Python interface a) more responsive to user input and b) not
require the GIL to be locked, enabling multithreading. Move more
of the simulation code into the python module code, stepping
manually and checking periodically for user interrupts.

TODO: currently I've broken the trace files. Improve this as
well, giving the user the option to either save the traces to
file or returning traces as Python objects. Ultimately, I don't
like how we create the SpikingChip with tracing options set -
since really this should be enabled when we start simulating.