pymodule.cpp: Support various trace objects
Better support traces from Python. Now the user can provide
either a filename, a file object, or request the data in memory
(as a Python object). The implementation is more flexible and
more Pythonic.

Additionally options allow you to skip writing the header (and
so allows you to append to existing trace files). This is
useful for various scripts.

TODO: change where files are opened in the SpikingChip class.
It doesn't really make sense to link file opening with object
creation - it should be when simulating that we provide trace
file objects.