Prevent builds in root project dir
Before I would by default recommend users setup their build
simply in the root directory. However, the cached CMake files
seem to cause problems across various builds. To fix this,
the general practice seems to be to require the creation of
a temporary build directory.

Throw a config error now if the user tries to 'make' in the
root dir.