Specify Python version & move PipelineUnit code
Allow user to specify the Python version to link. This
is useful for systems with many Python environments e.g.,
using conda. For some reason, occasionally the CMake env
doesn't pick up the right (active) version of Python. This
is probably the flexible way to do this.

Fix broken plugin linkage by moving PipelineUnit code from the
C++ file to the header i.e., make sure all public and protected
member functions for the PipelineUnit class are defined inline
in the header (otherwise we see missing symbols when using
plugins)

This is the simplest way to make sure the required code is
included within plugins, even if it means extra header code
and a bit of code duplication in the final binary.