Rework map_connections & support TrueNorth mask
Rather than passing in internal MappedConnection structs when
calling map_connections() for models [now renamed ->track_connections()]
just pass in the essential parameters i.e., the src/dest neuron IDs and
the synapse address. So far, this is all that is needed by any model,
and in the future, in theory any metadata can just be included along
with the edge attributes. It's hard to tell if we will require more internal
information for more complex synaptic models, but this approach simplifies
unit testing a lot, so roll with it for now.

Also add interface to set TrueNorth's random mask from an attribute. The
slight complication here is the attribute doesn't support unsigned integers
properly, but add some checks when passing the attribute to make sure the param
is unsigned. This will enable testing of this feature.