Fix Booksim config bug with wait_for_credits
For some unknown reason, if wait_for_tail_credit is set then Booksim
ignores the configured buffer queue size and instead defaults to
a buffer size of 1. This parameter shouldn't be needed, as all packets
have size of 1 flit i.e. the tail is also the head. When we remove this
config parameter, Booksim works and we see buffers correctly filled.