make_references.py: Include vol., issue and pages
For journal citations (including the main SANA-FE citation)
make sure to now include the full reference. In the future
we may need to handle preprints again, but for now make sure
to include the volume, issue (number) and page ranges.

Update the README.md and index.rst file with updated refs