[wip] Restore the subnet allocation and disable checks
Although this is a regression and undoes a bug fix where the subnet
starts at 1 instead of 0 (although this doesn't really matter),
we need to do this to be consistent with earlier experiments.
In future, we should also reset the subnet count every timestep.

Also disable multithreaded booksim checks for now.