Fix YAML parsing bug for neuron sim attributes
The bug caused simulator-specific neuron attributes to be parsed
incorrectly when given as a list of maps e.g.,

0: [log_spikes: true, log_potential:false]

In this case, the previous code would overwrite the entire neuron
configuration structure after each attribute, and so only the last
field would be set. The fix is to only optionally set attributes
that are encountered on each parser iteration.