chip.cpp: Function to clear chip's programmed SNN
Function to clear a mapped SNN from a h/w chip, making it possible
to reprogram a chip multiple times. This includes resetting state
and counters storing SNN connectivity e.g. axons and synaptic memory.

Make clearing overwriting the SNN an option that can be turned off/on,
in case a user wants to load() multiple SNNs onto the same chip
(probably a rarer use case). The default is now to overwrite on load()