Update latin squares and netlist fixes
Update Latin square generation to work with the latest version of
the simulator. This required some workarounds to ensure correct
neuron group naming and mapping. Add logic to handle the dummy
neuron allocation.

Some fixes to the netlist saving code, avoid duplication of default
attributes. Also exploring the noise generation model. This requires
a bit more work, but is working as before.