Fix bug saving YAML when file exists
When we aren't overwriting while saving a YAML, if that file
exists and isn't (valid) YAML, we were crashing out without
any helpful error. While this behavior is okay, make sure
to print a useful user message.

Also, as part of this, fix another bug with the build where
rapidYAML was *not* correctly configured to raise exceptions on
failures. Instead, rapidYAML was still aborting the program.
Move this configuration above the FetchProject to ensure correct
behavior.