pymodule.cpp: Optimize neuron group accesses
When accessing neurons within a neuron group, the previous
implementation would sequentially build a std::vector of references.
While it worked, it was very slow, especially for larger neuron groups.
Optimize this by creating a new "view" class which efficiently returns
only the requested neuron references. Now we have a O(1) access cost
again. This was verified on the nemo scripts.