indiveri.py: Clean up experimental plots
Now have the spiking digits raster plot cleaned up and stats
printed. The conclusion with this plot though is that perhaps
SHD is not the best app to demonstrate the Indiveri neurons,
as energy usage is dominated by synapse h/w.

Will now attempt to train the crossbar for spiking digits and
possibly copy this plotting over to the crossbar demo.