models.cpp: Use std::<optional> in accumulator
For AccumulatorWithDelay (which implements the timestep buffer
internally), change the accumulator values to be optionals.
If no inputs are received at a neuron, there should be no updates
made to that neuron (compartment). Previously, this dendrite h/w was
forcing every neuron to update every timestep.

Also register the "delay" attribute in the synapse model, to suppress an
annoying warning.