Pass timestep as update() param
As I'm documenting things, it seems much clearer/simpler to just
pass the simulation time-step as an initial function parameter rather
than maintaining some internal state in the base class object
(PipelineUnit). Doing so is more explicit and easier to explain to
the user. Unfortunately this is another (hopefully the last)
plugin API change.

Update unit tests to work with new interface (and to remove some
tests that no longer apply to the dendrite model)