pymodule.cpp: Fix bug where interrupts might hang
If an interrupt is sent from Python, either with open trace files or
after the sim has just completed, SANA-FE's python API would hang.
To fix: don't bother flushing if interrupting in the middle of a sim,
and check the Python state before printing, even if after the sim has
completed.